factorize the files(string|\Stringable ...$paths) method
