+++ docs/screenshot/MarkdownScreenshotBuilder.md
@@ -179,16 +179,12 @@ return $gotenberg
 ```
 
 ### files(Stringable|string ...\$paths)
-Add Markdown into a screenshot.<br /><br />Required to generate a screenshot from Markdown builder.<br />You can pass several files with that method.<br />
-
-> [!TIP]
-> See: [https://gotenberg.dev/docs/routes#screenshots-route](https://gotenberg.dev/docs/routes#screenshots-route)<br />
-> See: [https://gotenberg.dev/docs/routes#markdown-files-into-pdf-route ](https://gotenberg.dev/docs/routes#markdown-files-into-pdf-route )
+Adds files for the builder (overrides any previous files).<br />
 
 ```php
 return $gotenberg
     // Your builder call as ->html() and the rest of your configuration code
-    ->files('header.md','content.md','footer.md')
+    ->files('document.pdf', __DIR__'/../../public/document_2.pdf')
     ->generate()
     ->stream()
 ;
