+++ docs/screenshot/UrlScreenshotBuilder.md
@@ -67,6 +67,7 @@ class YourController
 ### Available methods
 
 - [downloadFrom](#downloadfromarray-downloadfrom)
+- [files](#filesstringablestring-paths)
 - [route](#routestring-name-array-parameters)
 - [url](#urlstring-url)
 - [addAsset](#addassetstringablestring-path)
@@ -117,6 +118,18 @@ Sets download from to download each entry (file) in parallel (URLs MUST return a
 return $gotenberg
     // Your builder call as ->html() and the rest of your configuration code
     ->downloadFrom([['url' => 'http://example.com/url/to/file', 'extraHttpHeaders' => ['MyHeader' => 'MyValue']], ['url' => 'http://example.com/url/to/file', 'extraHttpHeaders' => ['MyHeaderOne' => 'MyValue', 'MyHeaderTwo' => 'MyValue']]])
+    ->generate()
+    ->stream()
+;
+```
+
+### files(Stringable|string ...\$paths)
+Adds files for the builder (overrides any previous files).<br />
+
+```php
+return $gotenberg
+    // Your builder call as ->html() and the rest of your configuration code
+    ->files('document.pdf', __DIR__'/../../public/document_2.pdf')
     ->generate()
     ->stream()
 ;
