+++ src/Builder/Pdf/FlattenPdfBuilder.php
@@ -9,7 +9,6 @@ use Sensiolabs\GotenbergBundle\Builder\Behaviors\Dependencies\AssetBaseDirFormat
 use Sensiolabs\GotenbergBundle\Builder\Behaviors\DownloadFromTrait;
 use Sensiolabs\GotenbergBundle\Builder\Behaviors\WebhookTrait;
 use Sensiolabs\GotenbergBundle\Builder\Util\NormalizerFactory;
-use Sensiolabs\GotenbergBundle\Builder\Util\ValidatorFactory;
 use Sensiolabs\GotenbergBundle\Exception\MissingRequiredFieldException;
 
 /**
@@ -32,25 +31,13 @@ final class FlattenPdfBuilder extends AbstractBuilder
 
     public const ENDPOINT = '/forms/pdfengines/flatten';
 
-    /**
-     * If you provide multiple PDF files you will get ZIP folder containing all the converted PDF.
-     *
-     * @example files('document.pdf', __DIR__'/../../public/document_2.pdf')
-     */
-    public function files(string|\Stringable ...$paths): self
+    private const AVAILABLE_EXTENSIONS = [
+        'pdf',
+    ];
+
+    protected function getAllowedFilesExtensions(): array
     {
-        foreach ($paths as $path) {
-            $path = (string) $path;
-
-            $info = new \SplFileInfo($this->getAssetBaseDirFormatter()->resolve($path));
-            ValidatorFactory::filesExtension([$info], ['pdf']);
-
-            $files[$path] = $info;
-        }
-
-        $this->getBodyBag()->set('files', $files ?? null);
-
-        return $this;
+        return self::AVAILABLE_EXTENSIONS;
     }
 
     protected function getEndpoint(): string
