+++ src/Builder/Screenshot/MarkdownScreenshotBuilder.php
@@ -8,7 +8,6 @@ use Sensiolabs\GotenbergBundle\Builder\Attributes\WithBuilderConfiguration;
 use Sensiolabs\GotenbergBundle\Builder\Behaviors\ChromiumScreenshotTrait;
 use Sensiolabs\GotenbergBundle\Builder\BuilderAssetInterface;
 use Sensiolabs\GotenbergBundle\Builder\Util\NormalizerFactory;
-use Sensiolabs\GotenbergBundle\Builder\Util\ValidatorFactory;
 use Sensiolabs\GotenbergBundle\Enumeration\Part;
 use Sensiolabs\GotenbergBundle\Exception\MissingRequiredFieldException;
 use Sensiolabs\GotenbergBundle\Exception\PartRenderingException;
@@ -29,6 +28,10 @@ final class MarkdownScreenshotBuilder extends AbstractBuilder implements Builder
     }
 
     public const ENDPOINT = '/forms/chromium/screenshot/markdown';
+
+    private const AVAILABLE_EXTENSIONS = [
+        'md',
+    ];
 
     /**
      * The Twig file to convert into screenshot.
@@ -68,30 +71,9 @@ final class MarkdownScreenshotBuilder extends AbstractBuilder implements Builder
         return $this->contentFile($path);
     }
 
-    /**
-     * Add Markdown into a screenshot.
-     *
-     * Required to generate a screenshot from Markdown builder.
-     * You can pass several files with that method.
-     *
-     * @see https://gotenberg.dev/docs/routes#screenshots-route
-     * @see https://gotenberg.dev/docs/routes#markdown-files-into-pdf-route
-     *
-     * @example files('header.md','content.md','footer.md')
-     */
-    public function files(string|\Stringable ...$paths): self
+    protected function getAllowedFilesExtensions(): array
     {
-        foreach ($paths as $path) {
-            $path = (string) $path;
-            $info = new \SplFileInfo($this->getAssetBaseDirFormatter()->resolve($path));
-            ValidatorFactory::filesExtension([$info], ['md']);
-
-            $files[$path] = $info;
-        }
-
-        $this->getBodyBag()->set('files', $files ?? null);
-
-        return $this;
+        return self::AVAILABLE_EXTENSIONS;
     }
 
     protected function getEndpoint(): string
